+++ include/Armorial/Geometry/Rectangle/Rectangle.h
@@ -55,7 +55,7 @@ namespace Geometry {
          * \note This method will only work if the type T contains coordinates (.x() and .y() methods).
          */
         template<typename T>
-        [[nodiscard]] std::enable_if_t<Common::Types::has_coordinates_v<T>, bool> contains(const T& point, double margin) const {
+        [[nodiscard]] std::enable_if_t<Common::Types::has_coordinates_v<T>, bool> contains(const T& point, float margin) const {
             return (point.x() > (_topLeft.x() - margin) && point.x() < (_topRight.x() + margin))
                     && (point.y() > (_bottomLeft.y() - margin) && point.y() < (_topLeft.y() + margin));
         }
@@ -69,12 +69,12 @@ namespace Geometry {
         /*!
          * \return Returns this Rectangle instance width.
          */
-        [[nodiscard]] double width() const;
+        [[nodiscard]] float width() const;
 
         /*!
          * \return Returns this Rectangle instance height.
          */
-        [[nodiscard]] double height() const;
+        [[nodiscard]] float height() const;
 
         /*!
          * \return Returns a Vector2D instance containing the center of this Rectangle.
